불필요한 코드들 (분리를 완료하여서) 삭제, 다양한 컨텐트 타입을 처리하는 클래스 분리
